pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cosmic-clicker-pro"
version = "1.0.0"
description = "Multiplatform Clicker Game - Play on any device!"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["game", "clicker", "crossplatform", "mobile", "desktop", "web"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment",
]

dependencies = [
    "pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
android = [
    "buildozer>=1.5.0",
    "python-for-android>=2023.0.0",
]
web = [
    "pyodide>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/cosmic-clicker-pro"
Repository = "https://github.com/yourusername/cosmic-clicker-pro"
"Bug Reports" = "https://github.com/yourusername/cosmic-clicker-pro/issues"

[project.scripts]
cosmic-clicker = "src.cosmic_clicker_crossplatform:run_app"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]